Fixed papi_vector for infiniband_umad component.
The array of function pointers that the component defines must use the
naming convention papi_vector_t _x_vector where x is the name of the
component directory.

In this case, the name of the component directory is infiniband_umad
and not infiniband.

This change has not been tested yet due to OFED lib issues on our
local machines. There may be more changes required in order to get
the infiniband_umad component to work properly.