Make the lmsensors dynamically load the needed shared library
When attempting to build the current git repo of papi the build of the
files in the utils subdirectory failed because the lmsensors libraries
were not being linked in.  Rather than forcing the papi to link in the
lmsensor library during the build the lmsensors component has been
modified to dynamically load the needed libraries and enable the
lmsensors events when available.  This allows machines missing the
lmsensor libraries installed to still use papi.

Signed-off-by: William Cohen <wcohen@redhat.com>