Make the fallback generic_get_memory_info function more robust
On the aarch64 processor linux 4.11.0 kernels
/sys/devices/system/cpu/cpu0/cache is available, but the index[0-9]
subdirectories are not fully populated with information about cache
and line size, associativity, or number of sets.  These missing files
would cause the generic_get_memory_info function to attempt to read
data using a NULL file descriptor causing the program to crash.  Added
checks to see if every fopen was and fscan was successful and just say
there is no cache if there is any failure.