Use correct argument order for calloc function calls
Some calls to calloc in PAPI have the order of the arguments reversed.
According to the calloc man page the number of elements is the first
argument and the size of each element is the second argument.  Due to
alignment constraints the second argument might be rounded up.  Thus,
it is best not to not to swap the arguments to calloc.