Keep locks_pthreads test's amount of work reasonable on many core machines
The runtime of locks_pthreads test scaled by the number of processors
on the machine because of the serialized increment operation in the
test.  As more machines are available with 100+ processors the runtime
of locks_pthreads is becoming execessive.  Revised the test to specify
the approximate total number of iterations and split the work the
threads.