Ensure unique names for IBM Power 9 events
Older versions of PAPI use the event name to look up the libpfm event
number when doing the enumeration of the available events.  If there
were multiple events with the same name in libpfm, the earliest one
would be selected.  This selection would cause the enumeration of
events in papi_native_avail to get stuck looping on the first
duplicated named event in a pmu.  In the case of IBM Power 9 the
enumeration would get stuck on PM_CO0_BUSY. Gave each event a unique
name to avoid this unfortunate behavior.

Signed-off-by: 	William Cohen <wcohen@redhat.com>