add INTEL_X86_GRP_REQ umask flag for Intel X86
This patch adds a new umask flag for Intel X86 event tables.

It is used to handle the case where two umasks are dependent on each
other. For instance, if OPC_X can only be specified with UM1. Both
would be in separate groups. UM1 can be specified without OPC_X, but
not the opposite.

The flag is used in conjunction with the grpid field which is now
divided into two 8-bit numbers. The low-order bits are the classic
grpid, the upper 8 bits contain the required grpid.
code must use get_grpid() and get_req_grpid() accessor functions
for proper access to the field. If OPC_X is in grpid 2 and UM1 in
grpid 1, then the grpid for OPC_X would be: OPC_X.grpid = 1 << 8 | 2;

This feature is useful with some uncore filters.

Feature is supported for both intel_x86.c and snbep_unc.c.

Signed-off-by: Stephane Eranian <eranian@gmail.com>