add new PAPI_ECMP_DISABLED error
We can return this error if an event is added but the component involved
is disabled.

If a user moves working code to a system where perf_event_paranoid is
set to 3 (all perf events disabled) they will now get an error
indicating the component is disabled rather than an "event not found"
error which was confusing.