add INTEL_X86_FILT_UMASK umask flag for Intel X86
This flag is used to indicate that an event has a filter which must be
encoded in the umask of the event. Libpfm4 provides the encoding in
code[1] but the OS layer and possibly hardware expect it encoded
in the (extended) umask. There is a case in Skylake uncore UPI
PMU with the *_BASIC_HDR_MATCH event, for instance.

Signed-off-by: Stephane Eranian <eranian@gmail.com>