add INTEL_X86_FORCE_FILT0 umask flag for Intel x86
This patch adds a new umask flag on Intel x86. It is currently
only used for uncore PMUs. It indicates that the event has a
filter (filter0) and that it must be encoded even if its value is zero.
On some PMUs, the filter must be passed regardless of value.

Signed-off-by: Stephane Eranian <eranian@gmail.com>