Add Intel CascadeLake X core PMU support
CascadeLake X has the same core PMU as Intel Skylake X.
In fact, they share the same model number and differ
only in the steppings (CLX is stepping 5 and above).

This patch adds support by sharing the event table
with Skylake X. The detection code is custom because
it needs to take into account the stepping.

Signed-off-by: Stephane Eranian <eranian@gmail.com>