add get_num_events() support for Intel X86
This patch add support for the get_num_events() internal callback
for Intel X86 architecture. This is needed in case an event table
is shared by multiple PMU models and yet each PMU has some model-specific
events. This can happen between client and server models for instance.
In order to report the correct number of events available, we cannot
simply count the number of entries in the pe[] table. We need to
count how many are actually valid for a specific PMU model.

Signed-off-by: Stephane Eranian <eranian@gmail.com>