Change method for detecting available NVML component events
Previously PAPI nvml component used ROM version for detecting
the type of the GPU and find which events are supported. On
some newer cards, e.g. Tesla Kepler and Tesla Pascal, this gives
wrong result. Those card support GPU and memory utilization, for
example, but it was not detected as Kepler card may not have
powerROM, and PAPI nvml considers it as an old card.

So I changed the way the event availability is detected: just try
to obtain the info, and if it succeds, it is available.