Major addition: a component to access the rocm_smi library; this is the System Management Interface for AMD GPU devices. It allows monitoring of hardware elements; like power consumption, memory usage, PCIe throughput, fan speed, etc. It allows control for some hardware functions as well, via PAPI_write(), although these are untested (write requires root privileges to test). Included here are the component code, a tester for all readable events, and an incomplete tester for writing control values.
The tests are cpp; this is required for the AMD 'HIPP' compiler
to process an AMD Kernel that can exercise the GPU itself. The
rules and exports are bit complicated; for development the
rocm_smi_lib was installed and built in my user directory; in
production it would be in a system directory.