I fixed linux-rocm-smi.c to include an event per device called rocm_smi:::device=?:busy_percent; I overlooked this event in the first draft of the component.
I added a note to rocm_smi.h; we cannot use the distributed version
of this file; we have a compile error on one of the include files
<stdinit.h> that is not necessary; so we comment it out.

I created a Makefile for the rocm_smi/tests/ directory, it is just
a placeholder until we develop some standardized tests of the
rocm_smi component; but necessary to prevent an error during system
'make'.

I added rocm_command_line.out to the ROCM_SMI_Makefile. This is to
make non-standardized tests; and can be used as

make -f ROCM_SMI_Makefile <targetname>