linux-nvml.c is modified to accept an alternate name for the nvml library, which will default to the standard 'libnvidia-ml.so'. This is necessary on a system (Summit in particular) that doesn't have the standard link file to the current versioned lib. It will also provide flexibility for testing previous versions or new versions of the library. The library file name can be specified in Rules.nvml, as a compiler-line Define of NVML_LIBNAME.
Rules.nvml has comments added, one of which is an example of how to specify NVML_LIBNAME. Otherwise
it is unchanged; and the library name used will default to 'libnvidia-ml.so'.

README has been updated to describe this new capability, and for ICL staff contains examples of what
works on Summit.

PeakConfigure.sh had a typo that was corrected.