Added checks for improperly formatted lines in the user-provided event list.
If a line is missing a qualifier count, then it is discarded.

If a provided event name is either not available in the architecture or contains qualifiers, then the qualifier count is set to zero to prevent appending extraneous qualifiers, and the user is notified.

Also cleaned up string manipulation.

These changes were tested on the Intel Haswell architecture.