add PFMLIB_PMU_FL_DEPR flag
To mark a PMU model as deprecated. This is useful when a PMU model
is superseded by another one, yet the obsolete model must remain
for backward compatibility reason.

The ensures that a fully qualified event string with the old pmu
name will still be accepted. But when running on the matching
CPU model, the new PMU model will be selected by default when
the pmu model name is not specified.

Example: when running on pmu_old PMU model:
   - pmu_old::cycles is still accepted
   - pmu_new::cycles is accepted
   - cycles is mapped to pmu_new::cycles

Signed-off-by: Stephane Eranian <eranian@gmail.com>