add perf_events hw_smpl attribute
This patch adds a new attribute to perf_events OS support.
The attribute is called hw_smpl. It enables hardware sampling on an event.
Hardware sampling is CPU specific and therefore requires CPU specific code.
hw_smpl is a variation of precise sampling. It provides hardware assistance
to sample but does not guarantee precise attribution of samples to code.
With perf_events this is equivalent to setting attr.precise_ip = 1 which
is what this attribute does.

This patch only modifies the generic perf_event code.

Signed-off-by: Stephane Eranian <eranian@gmail.com>