Added check for whether or not the user provided a benchmark category.
When using the Counter Analysis Toolkit, if the user did not supply a benchmark category, then it will run the 'branch' benchmark by default and inform the user of such. The 'branch' benchmark executes the most quickly of all the categories, making it a suitable default.

These changes were tested on the Intel Haswell architecture.