---Correct cuda push/pop context consistency---
In _cuda_cleanup_eventset we attempt to push a current cuda context, set a new
cuda context to do some cleanup, then restore the original context with a pop.
(cuCtxPushCurrent, cuCtxPopCurrent). This was failing. We corrected it by doing
a Save+Restore instead of a Push+Pop using cuCtxGetCurrent, cuCtxSetCurrent,
different routines that do not require the cuda Context Stack, and have fewer
restrictions on their use.