fix handling of FETHR on Intel X86
This patch fixes several issues with the handling of the
Precise FRONTEND_RETIRED event on Intel X86 processors which
support it (Skylake and later). First, the FE latency field is
not 3 bits but 12. Second, the code was missing locked down
capability for the fe_thres modifier. Some events of the fe_thres
hardcoded, and therefore attempts to force a value should be
rejected. Third, when a umask does not hardcode a fe_thres
then the user can pass one. Note that not all umasks of the
event use the fe_thres.

Signed-off-by: Stephane Eranian <eranian@gmail.com>