add RETIRED_SSE_AVX_FLOPS event for AMD64 Fam17h Zen2
Was missing from initial commit. Added as PPR rev 0.54.

Note that this event by itself does not count correctly.
It needs large increment support, which means merging
of two consecutive counters. This is handled by the Linux
kernel starting with 5.6-rc4. The library simply encodes
the event as if it was like any other normal event.

Signed-off-by: Stephane Eranian <eranian@gmail.com>