update INST_RETIRED event encodings for Intel Icelake
This patch fixes several issues related to the encodings
of all the umasks of INST_RETIRED on Intel Icelake. Depending
on the umask, the event may have to be on fixed counter 1 and
this is expressed with a special encoding specific to perf_events.
Intel distinguishes PREC_DIST from ANY on fixed counter 1 vs.
ANY (architectural) on a generic counter.

Signed-off-by: Stephane Eranian <eranian@gmail.com>