lib/events/s390x_cpumf_events.h: Change counter name DFLT_CCERROR on s390
Change the counter name DLFT_CCERROR to DLFT_CCFINISH on IBM z15.
This counter counts completed DEFLATE instructions with exit code
0, 1 or 2. Since exit code 0 means success and exit code 1 or 2
indicate errors, change the counter name to avoid confusion.
This counter is incremented each time the DEFLATE instruction
completed regardless if an error was detected or not.

This change is in sync with kernel
commit 3d3af181d370 ("s390/cpum_cf,perf: change DFLT_CCERROR counter name")

Signed-off-by: Thomas Richter <tmricht@linux.ibm.com>