Add string length check before strncpy() and strcat() calls in _mx_init_component()
Myrinet Express-related component MX modules are initialized with the _mx_init_component() function,
which is called from the PAPI_library_init() function.
The popen(3) call runs a loadable module called "mx_counters",
and if the loadable module does not exist,
it attempts to run a loadable module called "./components/mx/utils/fake_mx_counters".
In an environment where there are no "mx_counters" and "./components/mx/utils/fake_mx_counters" loadable modules,
popen(3) will be called twice uselessly.
popen(3) internally calls pipe(2) once, fork(2) twice and exec(2) once.

The size of the user space of the application calling the PAPI_library_init() function affects the performance of fork(2),
which is called as an extension of popen(3).
As a result, the performance of the PAPI_library_init() function is affected by the amount of user space in the application
that called the PAPI_library_init() function.

In the _mx_init_component() function,
the MX module only needs to be able to verify that a load module named "mx_counters" exists.
We improved the _mx_init_component() function to call fopen(3) instead of popen(3).
We add string length check before strncpy() and strcat() calls in _mx_init_component() function.