Sysdetect: system information detection component
The SYSDETECT component allows PAPI users to query comprehensive system
information. The information is gathered at PAPI_library_init() time and
presented to the user through appropriate APIs. The component works
similarly to other components, which means that hardware information for
a specific device might not be available at runtime if, e.g., the device
runtime software is not installed.

At the moment the infrastructure defines the following device types:

  - PAPI_DEV_TYPE_ID__CPU        : for all CPU devices from any vendor
  - PAPI_DEV_TYPE_ID__NVIDIA_GPU : for all GPU devices from NVIDIA
  - PAPI_DEV_TYPE_ID__AMD_GPU    : for all GPU devices from AMD

Every device is scanned to gather information when the component is
initialized. If there is no installed hardware for the considered
device type, is not found, the corresponding information is filled with
zeros.

This patch also adds a new utility program, called papi_hardware_avail,
that prints out to the command line what hardware is installed for each
type and the specifications of each device.