linux-memory: change cache parsing so it works on ARM servers
On Linux we parse files under /sys/devices/system/cpu/ to determine
the various cache settings.  The old code assumes certain files, such
as associativity and linesize are always there (because they are on x86).

This updated code instead of exiting with an error if the files don't
exist sets the values to 0.  This allows the cache values to be returned
on ARM systems such as Ampere servers.

This could potentially break user code if they are taking the cache
values and doing division (such as taking the cache size and dividing
by the linesize: if linesize is zero they could get a divide-by-zero error).
I'm not sure if there's a way around this without redesigning how
the meminfo structure works.