RocmCmp: fix bug in event reporting while running papi_component_avail
In the papi_component_avail() utility program the component private init
function is called multiple times. The first time the component events
are initialized as expected and the disable flag is set to the error
code returned while performing the process, along with the reason the
component might be disabled. The second time the private init function
of the component is called is when trying listing the supported events.
In this case the init does not remember the error code previously
returned and sets the disabled flag to PAPI_OK instead. This same bug
was already fixed by Tony Castaldo in patch e646d570 for the cuda
component.