sysdetect/rocm: fix hsa_status_string arguments
hsa_status_string takes an hsa_status_t code and returns a pointer to
const char with the error message associated to the status code. In the
code we were passing a pointer to a status array of chars rather than a
pointer to a const char. This is fixed by this patch.