sysdetect/rocm: explicitly look for rocm in PAPI_ROCM_ROOT
Similarly to the ROCm component sysdetect also requires PAPI_ROCM_ROOT
to be defined so that the user is explicitly forced to define where in
the file system tree the ROCm installation to be used is located. This
prevents from situations in which the user believes she is using a
certain ROCm version while in reality the component is picking up the
system environment defined one.