cuda: extend device counting with /proc filesystem
/sys based device counting for cuda relies on the linux display
rendering manager populating the corresponding entries in the
filesystem. This is not always the case and depends on the specific
linux configuration. Thus, this method might cause the component wrongly
detecting no cuda devices in a system that has some. Another source of
information for cuda devices is /proc file system. This patch extends
the current /sys functionality with /proc information.