cuda: fix overcounting of cuda devices
Cuda device counting is done by scanning the file system for the related
device number information. There are two places in the file system where
this information can be found: in /sys and in /proc. The second being a
fallback in case the first does not contain the desired information. The
/proc based device counting goes through all the directories in the
/proc/driver/nvidia/gpus, including '.' and '..', thus overcounting the
number of devices. This patch fixes the problem by filtering such
directories.