sysdetect: fix infinite loop bug
When there is no node is /sys/devices/system/cpu/ sysdetect will loop
indefinitely looking for the node affine to the thread specified.
Instead we should look for existence of node0 for cpu0. If this is
present then other nodes will likely to be in the file system tree.
Otherwise there is no point looking further and we just assume there is
only one numa node.