PAPI_get_hardware_info: improve PAPI_hw_info_t for ARM processors
Currently, it is not possible to determine which company the ARM processor
was designed by from the PAPI_hw_info_t available in PAPI_get_hardware_info().
On ARM processors, the PAPI_hw_info_t obtained with PAPI_get_hardware_info()
does not contain information indicating which company was designed.
For ARM processors, improve the vendor and vendor_string entries
in PAPI_hw_info_t, which can be retrieved with PAPI_get_hardware_info(),
to include information indicating which company was designed.

Signed-off-by: Masahiko, Yamada <yamada.masahiko@fujitsu.com>