high-level: use variable to select between single and multi-thread mode
Currently, the HL API always assumes multi-thread mode. This means that
multiple threads in the program can create and manage PAPI event sets.
This is not a valid assumption as the behavior is different for
single-thread monitoring programs. This patch, introduces a new
environment variable named PAPI_HL_THREAD_MULTIPLE that allows to select
single and multi-thread mode explicitly in the HL API. To avoid
affecting existing applications and tests the default is multi-thread
monitoring. If the variable is set to "0" single-thread mode is selected
instead. For explicitly setting multi-thread monitoring the variable has
to be set to "1".