sysdetect: extend PAPI with system detection and querying APIs
Queries performed by accessing PAPI internal data structures are not
easily maintainable. Once an internal data structure is exposed to users
these rely on it not being changed and ties our hands from an
implementation standpoint. This patch introduces a new set of APIs that
can be used to query system attributes for different devices through the
system detection component. The APIs are generic enough to allow
extending the capabilities with new hardware devices, as they become
available, and allow to separate user interfaces from the implementation
of the underlying functionality.

Because the new APIs have to be always functional we configure the
sysdetect component by default and initialize it lazily, i.e., only when
the corresponding APIs are called by the user.