Issue102: Remove CUDA11 callback subscriber
The CUPTI callback subscriber introduced to monitor contexts created
a problem for packages that use PAPI for CUDA performance counter
collection.  Specifically, it prevented registering of custom
profiling/tracing callbacks, resulting in
`CUPTI_ERROR_MULTIPLE_SUBSCRIBERS_NOT_SUPPORTED` when attempting to
register one.

This changeset is effectively a targeted reversion of the callback
subscriber logic from the commit that introduced it.  Commit that
introduced subscriber: 9ff1d73dae9a7b297a54a77fac5fdb3957041452

With this update in place, the `cuda` component context capturing
behavior is consistent between the legacy and update CUPTI11 version
of the code.  Requiring that applications create and set the context
used to run the kernels prior to calling `PAPI_add_events()`.