powercap: fix event lookup in _powercap_read()
The function read_powercap_value() should be given the index of the powercap event, not the position of that event in the event set.
The event-set index maps to the powercap-event index via the 'which_counter' array, which is already used in the function _powercap_write().

These changes were tested on the Intel Skylake and Cascade Lake architectures.