CUDA11 Profiler Active Context Sensitivity
The CUPTI11 portion of the `cuda` component has shown sensitivities
for a calling threads active context when using CUDA11 versions < 11.2.
Specifically, it was found that not pushing the session context onto the
stack prior to calling `cuptiProfilerSetConfig` would result in a failure.

This change set addresses this issue by leveraging the same mechanics that
are used when calling other CUPTI routines, namely pushing the session
context onto the stack and popping it off once we are done with it.