sysdetect: modify configure.in logic to parse '--with-CPU=x86'
When the configuration is invoked with "--with-CPU=x86", it should add 'x86_cpuid_info.c' to MISCSRCS.

When the configuration is invoked on an architecture of the x86_64 family, and "--with-CPU=x86" is not specified, the build will proceed normally.
However, the build will fail on x86_64 when "--with-CPU=x86" is specified because this flag bypasses the check for the x86_64 family but does not add 'x86_cpuid_info.c' to MISCSRCS.
Thus, if "--with-CPU=x86" is specified, "x86" must be included in the list of CPUs for which 'x86_cpuid_info.c' is added as a source file.

These changes have been tested on the Intel Cascade Lake and Skylake architectures.