rocm: fix assign eventset to component
PAPI_assign_eventset_component() assigns an eventset to a component of
certain index. The function relies on component information to allocate
data structures for the component. One such parameter is num_mpx_cntrs.
The component was setting this to -1 in rocm_init_component() causing
any malloc to fail. Additionally, when rocm_init_private() is finally
called, it also resets num_mpx_cntrs to the number of native events
detected for the device. This is wrong as the framework relies on this
parameter when freeing allocated data structures, e.g., EventInfoArray.