ctests: allow access to PAPI_EDELAY_INIT components
Some of the tests, such as all_native_events and get_event_component,
check the 'disabled' flag of the component before accessing it. Device
components, such as cuda and rocm, set the 'disabled' flag to
PAPI_EDELAY_INIT, which signifies the component is a delayed
initialization one. Thus, the event table in the component is
initialized only when events are accessed (e.g. PAPI_enum_cmp_event).