perf_event_u: replace PAPI_ENOCMP with PAPI_ECMP
PAPI_ENOCMP should be used to indicate that the requested component is
not available in the component index (e.g. because it wasn't
initialized). PAPI_ECMP, on the other hand, should be used when the
component is initialized but some requested feature is not supported by
the component (e.g. the component is not compatible with the feature).
By its own definition no component can return PAPI_ENOCMP.