cat: vector FLOPs benchmark for non-x86 architectures bug fix
The driver code for the vector benchmark could not call the functions
for the vector FLOPs kernels because they were declared 'static'.
For builds which use either the NEON or ALTIVEC intrinsics, these
static functions are now wrapped, so they can be called by the driver.

These changes have been tested on the IBM POWER9 architecture.