sysdetect: fix cache info data structure name
With commit number 9f8e6b0 the sysdetect data structures, that were
originally hosted in papi.h, are moved to the sysdetect.h instead. To
account for this, the data structures prefix was changed from PAPI_ to
_sysdetect_. This change, however, was erroneously skipped for the arm
files. This patch fixes this problem reflecting the name change in arm
files.