Return PAPI library version even if PAPI is not initialized
Currently, there is no way for the user to compare his version of the
PAPI library with the version of the library being loaded at runtime.
PAPI_library_init() takes the version of the user library and compares
it with the version of the library loaded. If the two don't match, it
returns an error.

PAPI_get_opt() also provides access to some library information, like
the version, but this only works if PAPI has been correctly initialized.
This patch extends PAPI_get_opt() to provide the library version
regardless PAPI being initialized at all.