sysdetect: fix problem with missing shutdown_thread implementation
Originally, init_thread and shutdown_thread were not implemented in the
sysdetect component. However, this causes issues when
PAPI_register_thread and PAPI_unregister_thread are used. In the case of
unregister, the framework will go through all the enabled components and
call shutdown_thread for each of them. Since the sysdetect did not
implement these functions a default (PAPI_ECMP) error would be returned.

This patch adds the missing functions to sysdetect.

Solves issue #116