sysdetect: fix warning in papi_fwrappers.c
papi_fwrappers.c is used to generate multiple wrapper versions for
fortran. Because of a global variable not declared static the different
versions cause a redefinition of the symbols when used with recent
versions of the gcc compiler (as the compiler does link time
optimizations). Declaring the variable static should fix the problem.