rocm: SQ_WAVES does not reflect logical waves
SQ_WAVES counts the number of logical waves, plus the waves that are
restored due to context switching. This patch computes the logical
number of waves as SQ_WAVES - SQ_WAVES_RESTORED. For those architectures
that do not support SQ_WAVES_RESTORED (preceeding MI200) the tests
return with a warning and the number of waves check is ignored.