Use fgets in place of fscanf functions to avoid possible buffer overflows
There were several locations in the PAPI code that used fscanf calls
like the following statement to read in information:

               result=fscanf(fff,"%s",allocation_policy_string);

The problem with this statement is that the fscanf could possibly write past
the end of allocation_policy_string.  To limit the write to the size of the
allocation_policy_string an fgets like the following is used in its place:

               str_result=fgets(allocation_policy_string, BUFSIZ, fff);

One set of fscanf were for the generic memory information code reading
the cache characteristics.  Another fscanf was in the sysdetect
component reading of cache characteristics.