pcp: warning instead of error when 'reason' string truncated
When the hostname is too long, there is not enough memory allocated for the error 'reason' string.
This caused the component to prematurely exit initialization when the PM daemon is not active.
Instead, a warning is now issued, and the initialization exits appropriately.

Additionally, the size of the 'reason' string has been increased to accommodate longer host names.

These changes have been tested on the IBM POWER9 architecture.