sysdetect: hook mpi tests to NO_MPI_TESTS
The configure step in PAPI checks whether MPI tests can be enabled or
not. If not it sets NO_MPI_TESTS to yes. This variable is then used in
ctests/Makefile.recipies to enable or disable MPI tests. The sysdetect
tests were not relying on this variable. Instead sysdetect relied on
MPICC being set which is no accurate. This patch make the MPI checks
more uniform across the code by adding NO_MPI_TESTS checks in sysdetect
tests too.