Improved the branch misprediction validation test.
The previous version of the branch misprediction validation test relied on
the libc function random() to generate entropy. However, this function
introduced 15x more branches than the number of branches in the code of the
validation test, polluting the results.
The new code uses an inline Xorshift pseudo-random number generator
which is more than sufficient to confuse the branch predictor, and
does not contain any branch instructions so it does not pollute the
event measurement. Also, the logic of the test has been simplified.