sysdetect: fix compile error with gcc 10
Some of the symbols used by the sysdetect component clash with rocm
and cuda components as they are not defined static. This problem only
seems to affect latest versions of the gcc compiler (>= 10). This is
due to how gcc places global variables that do not have an initializer
(tentative definition variables in the C standard). These variables are
places in the .BSS section of the object file. This avoids the merging
of tentative definition variables by the linker, which causes multiple
definition errors. (This happens by default in gcc and corresponds to
the -fno-common option).