cuda: fix compile error with gcc 10
Some of the symbols used by the cuda component clash with the nvml
component as they are not defined static. This problem only affects
latest versions of the gcc compiler (>= 10). This is due to how gcc
places global variables that do not have an initializer (tentative
definition variables in the C standard). These variables are placed
in the .BSS section of the object file. This avoids the merging of
tentative definition variables by the linker, which causes multiple
definition errors. (This happens by default in gcc and corresponds
to the -fno-common option).