rocm: refactor rocp_ctx_read (2/28)
The events_id array, containing the id of the events requested by the
user, is passed to rocp_ctx_open and can be saved in the rocp_ctx
returned by this function. It is not necessary to pass the array again
as argument to rocp_ctx_read. Thus, remove it from the argument list of
rocp_ctx_read.