rocm: refactor event collision detection (5/28)
Rename compare_events to verify_events to indicate the change in
functionality for the function. Used to compare events, returning either
0 if they all matched or an integer if the did not. Now verify_events
returns PAPI_OK if the events match and PAPI_ECNFLT if they don't.
verify_events also checks whether the rocprofiler callbacks are set or
not. If not the function exits immediately as there cannot be any event
conflict.