rocm: refactor rocp_ctx_open function (9/28)
The goal of this patch is to remove the need for passing the event table
reference to the rocp_ctx_open function. To avoid too many changes in
the code here we introduce a new rocp_ctx_open_v2 function and replace
rocp_ctx_open with it in a later commit.