rocm: refactor event duplication verification logic (18/28)
Currently, update_native_events removes event duplication whenever
that happens. This is the case for events that differ by instance
number. The rocp layer reports events instances as separate native
events. The logic to remove duplicate events is messy and probably
not even needed as the user will normally add events without
indicating the instance of the event. This patch removes such logic.