powercap-ppc: fix warning in strncpy
strncpy causes the following warning in linux-powercap-ppc.c:

warning: '__builtin_strncpy' specified bound 1024 equals destination size
   62 |     char *retval = strncpy( dst, src, size );
      |                    ^~~~~~~

the problem is that size is the same size as dst. If src is also the
same length, the null termination character will not be copied over.
Instead copy only size - 1 and terminate the string manually.