rocm: fix bug in sampling_ctx_open
The function creates one profiling context per device. The way the agent
corresponding to the device is selected was erroneous. This caused
different threads monitoring different devices with different eventsets
to all access the counters from the first device. The fix is not to
select the agent using a for loop index but instead to use that index to
get the device id from the devs_id array.