nvml: fix support for multiple devices
Replace each cudaGetDevicePtr() call with a table lookup. This tracks the correct device ID when counting events; whereas, cudaGetDevicePtr() will only ever return a single ID with the way it was used.

The dependency on CUDA unnecessarily restricts multi-process jobs on Summit. Removing this dependency was necessary to properly support multiple devices.

These changes were tested on the Summit supercomputer, which contains the IBM POWER9 and NVIDIA Tesla V100 architectures.