cat: non-core events in multithreaded benchmarks
Each thread in the data-cache benchmarks previously added events to its
local event set.
This caused an error for events that are not from the perf_event (core)
component.

To accommodate such events, we check for the component to which an event
belongs.
If the event does not belong to the core component, then only the thread
with ID 0 adds the event to its event set.

These changes have been tested on the AMD Zen3 architecture.