cat: allow user to specify DCR/DCW sampling
To provide more flexible benchmarks, we allow the user to specify the
number of measurements for each level of the memory hierarchy.

In addition, we include user-definable parameters to accommodate the
different cache levels shared by different numbers of cores.

These changes have been tested on the AMD Zen3 and Zen4 architectures.