buildsystem: fix install target in Makefile
PR #464 introduced a --disable-fortran flag that allows users to disable
fortran header and wrappers generation in case the user does not need
them. Commit 40b7afc also introduces a bug as the fortran headers no
longer generated by the configure are still part of the install target.
This patch fixes the problem.