rocm: user HIP_PATH for hipcc compiler in tests Makefile
The rocm tests assume hipcc is located under the same root directory as
the rest of rocm toolkit software. Spack installs rocm dependencies in
separate directories however, which breaks this assumption. This patch
introduces a HIP_PATH variable that, if unset, is set automatically to
PAPI_ROCM_ROOT. Spack can use this variable to let the tests Makefile in
the PAPI rocm component know where the hipcc compiler is located.