Ensure strings copied in _coretemp_init_component function NULL are terminated
The strncpy function will not place a NULL character at the end of the
string if the string being copied is the same length or longer than
the destination of the strncpy function.  Switching the code in the
_coretemp_init_component function to use snprintf and checking the
return value of snprintf to verify the copied string fits in the
destination.