sde_lib: Added reference counts for proper unregistering of groups.
Counters can belong in groups, even multiple groups, and groups can recursively
belong in larger groups. This means that a counter (or group) cannot be
unregistered and freed without keep track of which groups it belong to.
Now each counter has a reference counter "ref_count" which is incremented
when it's added in a group and decremented when the counter is unregistered,
or when a parent group is unregistered.