rocm: move extern declarations to config header
The rocm lock and the profiling mode variables need to be shared between
the front-end and the back-end. The reason for the lock is that this has
to be initialized by the front-end which is the only one with access to
the required information. This lock design in PAPI is flawed as it is
hard to extend.