powercap: test counter read permissions
Check that the files inside /sys/class/powercap/intel-rapl:<n>
directories not only exist, but are readable.  On recent Linux kernels,
"energy_uj" is by default readable by root only, which is something that
PAPI fails to detect, resulting in 0 being returned for that counter
without any indication of a problem.